Move kotlin lib onboardingpayload to package matter
