+++ examples/android/CHIPTool/app/src/main/java/com/google/chip/chiptool/CHIPToolActivity.kt
@@ -28,9 +28,6 @@ import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
-import chip.onboardingpayload.OnboardingPayload
-import chip.onboardingpayload.OnboardingPayloadParser
-import chip.onboardingpayload.UnrecognizedQrCodeException
 import com.google.chip.chiptool.provisioning.AddressCommissioningFragment
 import com.google.chip.chiptool.provisioning.DeviceProvisioningFragment
 import com.google.chip.chiptool.provisioning.EnterNetworkFragment
@@ -40,6 +37,9 @@ import com.google.chip.chiptool.setuppayloadscanner.CHIPDeviceDetailsFragment
 import com.google.chip.chiptool.setuppayloadscanner.CHIPDeviceInfo
 import com.google.chip.chiptool.setuppayloadscanner.CHIPLedgerDetailsFragment
 import com.google.chip.chiptool.util.DeviceIdUtil
+import matter.onboardingpayload.OnboardingPayload
+import matter.onboardingpayload.OnboardingPayloadParser
+import matter.onboardingpayload.UnrecognizedQrCodeException
 import org.json.JSONObject
 
 class CHIPToolActivity :
