+++ examples/tv-app/android/App/platform-app/src/main/java/com/matter/tv/server/fragments/QrCodeFragment.java
@@ -9,10 +9,10 @@ import android.widget.ImageView;
 import android.widget.SeekBar;
 import android.widget.TextView;
 import androidx.fragment.app.Fragment;
-import chip.onboardingpayload.DiscoveryCapability;
-import chip.onboardingpayload.OnboardingPayload;
-import chip.onboardingpayload.OnboardingPayloadException;
-import chip.onboardingpayload.OnboardingPayloadParser;
+import matter.onboardingpayload.DiscoveryCapability;
+import matter.onboardingpayload.OnboardingPayload;
+import matter.onboardingpayload.OnboardingPayloadException;
+import matter.onboardingpayload.OnboardingPayloadParser;
 import com.matter.tv.server.QRUtils;
 import com.matter.tv.server.R;
 import com.matter.tv.server.service.MatterServant;
