+++ src/controller/java/BUILD.gn
@@ -272,20 +272,20 @@ kotlin_library("onboarding_payload") {
   deps = [ ":tlv" ]
 
   sources = [
-    "src/chip/onboardingpayload/Base38.kt",
-    "src/chip/onboardingpayload/CommissioningFlow.kt",
-    "src/chip/onboardingpayload/DiscoveryCapability.kt",
-    "src/chip/onboardingpayload/ManualOnboardingPayloadGenerator.kt",
-    "src/chip/onboardingpayload/ManualOnboardingPayloadParser.kt",
-    "src/chip/onboardingpayload/OnboardingPayload.kt",
-    "src/chip/onboardingpayload/OnboardingPayloadParser.kt",
-    "src/chip/onboardingpayload/OptionalQRCodeInfo.kt",
-    "src/chip/onboardingpayload/QRCodeBasicOnboardingPayloadGenerator.kt",
-    "src/chip/onboardingpayload/QRCodeOnboardingPayloadGenerator.kt",
-    "src/chip/onboardingpayload/QRCodeOnboardingPayloadParser.kt",
-    "src/chip/onboardingpayload/VendorId.kt",
-    "src/chip/onboardingpayload/Verhoeff.kt",
-    "src/chip/onboardingpayload/Verhoeff10.kt",
+    "src/matter/onboardingpayload/Base38.kt",
+    "src/matter/onboardingpayload/CommissioningFlow.kt",
+    "src/matter/onboardingpayload/DiscoveryCapability.kt",
+    "src/matter/onboardingpayload/ManualOnboardingPayloadGenerator.kt",
+    "src/matter/onboardingpayload/ManualOnboardingPayloadParser.kt",
+    "src/matter/onboardingpayload/OnboardingPayload.kt",
+    "src/matter/onboardingpayload/OnboardingPayloadParser.kt",
+    "src/matter/onboardingpayload/OptionalQRCodeInfo.kt",
+    "src/matter/onboardingpayload/QRCodeBasicOnboardingPayloadGenerator.kt",
+    "src/matter/onboardingpayload/QRCodeOnboardingPayloadGenerator.kt",
+    "src/matter/onboardingpayload/QRCodeOnboardingPayloadParser.kt",
+    "src/matter/onboardingpayload/VendorId.kt",
+    "src/matter/onboardingpayload/Verhoeff.kt",
+    "src/matter/onboardingpayload/Verhoeff10.kt",
   ]
 }
 
@@ -298,7 +298,7 @@ kotlin_library("onboardingpayload_manual_code_test") {
     "${chip_root}/third_party/java_deps:truth",
   ]
 
-  sources = [ "tests/chip/onboardingpayload/ManualCodeTest.kt" ]
+  sources = [ "tests/matter/onboardingpayload/ManualCodeTest.kt" ]
 
   kotlinc_flags = [ "-Xlint:deprecation" ]
 }
@@ -312,7 +312,7 @@ kotlin_library("onboardingpayload_qr_code_test") {
     "${chip_root}/third_party/java_deps:truth",
   ]
 
-  sources = [ "tests/chip/onboardingpayload/QRCodeTest.kt" ]
+  sources = [ "tests/matter/onboardingpayload/QRCodeTest.kt" ]
 
   kotlinc_flags = [ "-Xlint:deprecation" ]
 }
