+++ src/controller/java/src/matter/onboardingpayload/Base38.kt
@@ -15,7 +15,7 @@
  *    limitations under the License.
  */
 
-package chip.onboardingpayload
+package matter.onboardingpayload
 
 private val kCodes =
   charArrayOf(
