+++ src/controller/java/src/matter/onboardingpayload/DiscoveryCapability.kt
@@ -15,14 +15,11 @@
  *    limitations under the License.
  */
 
-package chip.onboardingpayload
+package matter.onboardingpayload
 
-/**
- * Enum values for possible flows for out-of-box commissioning that a Matter device manufacturer may
- * select for a given product.
- */
-enum class CommissioningFlow(val value: Int) {
-  STANDARD(0), // Device automatically enters pairing mode upon power-up
-  USER_ACTION_REQUIRED(1), // Device requires a user interaction to enter pairing mode
-  CUSTOM(2) // Commissioning steps should be retrieved from the distributed compliance ledger
+/** Enum values for possible bits in the onboarding paylod's discovery capabilities bitmask. */
+enum class DiscoveryCapability {
+  SOFT_AP,
+  BLE,
+  ON_NETWORK
 }
