+++ src/controller/java/src/matter/onboardingpayload/OptionalQRCodeInfo.kt
@@ -15,11 +15,26 @@
  *    limitations under the License.
  */
 
-package chip.onboardingpayload
+package matter.onboardingpayload
 
-/** Enum values for possible bits in the onboarding paylod's discovery capabilities bitmask. */
-enum class DiscoveryCapability {
-  SOFT_AP,
-  BLE,
-  ON_NETWORK
+enum class OptionalQRCodeInfoType {
+  TYPE_UNKNOWN,
+  TYPE_STRING,
+  TYPE_INT32,
+  TYPE_INT64,
+  TYPE_UINT32,
+  TYPE_UINT64
+}
+
+open class OptionalQRCodeInfo {
+  var tag: Int = 0
+  var type: OptionalQRCodeInfoType = OptionalQRCodeInfoType.TYPE_UNKNOWN
+  var data: String? = null
+  var int32: Int = 0
+}
+
+class OptionalQRCodeInfoExtension : OptionalQRCodeInfo() {
+  var int64: Long = 0
+  var uint32: Long = 0
+  var uint64: Long = 0
 }
