+++ src/controller/java/src/matter/onboardingpayload/QRCodeOnboardingPayloadParser.kt
@@ -15,7 +15,7 @@
  *    limitations under the License.
  */
 
-package chip.onboardingpayload
+package matter.onboardingpayload
 
 import java.util.concurrent.atomic.AtomicInteger
